Unify bootstrap logic and introduce Runner
