+++ Sources/Runner/main.swift
@@ -1,0 +1,9 @@
+import PointFree
+
+// Bootstrap
+
+_ = try! PointFree
+  .bootstrap()
+  .run
+  .perform()
+  .unwrap()
